beam/requirements.txt
absl-py==0.6.1
apache-beam==2.8.0
Pillow==5.3.0
tensorflow==1.12.0
